Refactored integration tests
